Bruh where
