+++ web/app.fluidity.money/app/repositories/README.md
@@ -1,0 +1,25 @@
+# Repositories
+
+I couldn't think of a better name for these.
+
+---
+
+### What's a repository
+
+These process queries into a more useful format, like a serializer;
+
+Transform data to a simpler more useful format;
+
+Then store things in memory as a useState.
+
+Why not work with these directly in the query?
+
+We _did_ and it was:
+
+- Hard to test
+- Easy to duplicate code
+- Easy to introduce bugs
+- Cluttered up page code
+- Hard to maintain
+
+That's why we have this 😄
