+++ web/app.fluidity.money/app/routes/$network/dashboard/rewards/unclaimed.tsx
@@ -4,13 +4,11 @@ import type { IRow } from "~/components/Table";
 
 import config from "~/webapp.config.server";
 import { LoaderFunction, json } from "@remix-run/node";
-import { captureException } from "@sentry/react";
 import { useLoaderData, useLocation } from "@remix-run/react";
-import { useState, useEffect, useContext } from "react";
+import { useState, useContext } from "react";
 import FluidityFacadeContext from "contexts/FluidityFacade";
 import { UserRewards } from "./common";
 import { getTxExplorerLink, getAddressExplorerLink } from "~/util";
-import { useUserUnclaimedRewards, useUserRewardsByAddress } from "~/queries";
 
 import { motion } from "framer-motion";
 import { Table } from "~/components";
@@ -22,6 +20,11 @@ import {
 } from "@fluidity-money/surfing";
 import useViewport from "~/hooks/useViewport";
 import { getBlockExplorerLink } from "~/util/chainUtils/links";
+import {
+  getUnclaimedRewards,
+  TokenUnclaimedReward,
+} from "~/repositories/getUnclaimedRewards";
+import { getClaimedRewards } from "~/repositories/getClaimedRewards";
 
 export const unstable_shouldReload = () => false;
 
@@ -70,31 +73,15 @@ type LoaderData = {
   network: Chain;
 };
 
-type TokenUnclaimedReward = {
-  symbol: string;
-  reward: number;
-};
-
 const UnclaimedWinnings = () => {
   const { network, fluidTokenMap, tokenDetailsMap } =
     useLoaderData<LoaderData>();
 
-  const { connected, address } = useContext(FluidityFacadeContext);
+  const { address } = useContext(FluidityFacadeContext);
 
-  const [
-    { userUnclaimedRewards, unclaimedTxs, unclaimedTokens },
-    setUnclaimedRewardsRes,
-  ] = useState<{
-    unclaimedTxs: UserUnclaimedReward[];
-    unclaimedTokens: TokenUnclaimedReward[];
-    userUnclaimedRewards: number;
-  }>({
-    unclaimedTxs: [],
-    unclaimedTokens: [],
-    userUnclaimedRewards: 0,
-  });
-
-  const [userTotalRewards, setUserTotalRewards] = useState(0);
+  const { userUnclaimedRewards, unclaimedTxs, unclaimedTokens } =
+    getUnclaimedRewards(address ?? "", network);
+  const userTotalRewards = getClaimedRewards(address ?? "", network);
 
   const [{ networkFee, gasFee }] = useState({
     networkFee: 0,
@@ -112,99 +99,7 @@ const UnclaimedWinnings = () => {
   const _pageUnsafe = _pageStr ? parseInt(_pageStr) : 1;
   const txTablePage = _pageUnsafe > 0 ? _pageUnsafe : 1;
 
-  useEffect(() => {
-    if (!connected || !address) return;
-
-    // Get Unclaimed Rewards - Expect to fail if Solana
-    (async () => {
-      try {
-        const { data, error } = await useUserUnclaimedRewards(
-          network,
-          address ?? ""
-        );
-
-        if (!data || error) return;
-
-        const { ethereum_pending_winners: rewards } = data;
-
-        const userUnclaimedRewards = rewards.reduce((sum, transaction) => {
-          const { win_amount, token_decimals } = transaction;
-
-          const decimals = 10 ** token_decimals;
-          return sum + win_amount / decimals;
-        }, 0);
-
-        const unclaimedTokens = Object.entries(
-          rewards.reduce((map, transaction) => {
-            const { win_amount, token_decimals, token_short_name } =
-              transaction;
-            const reward =
-              (map[token_short_name] ?? 0) + win_amount / 10 ** token_decimals;
-
-            return {
-              ...map,
-              [token_short_name]: reward,
-            };
-          }, {} as { [tokenName: string]: number })
-        ).map(([symbol, reward]) => ({ symbol, reward }));
-        
-        console.table(rewards);
-        setUnclaimedRewardsRes({
-          unclaimedTxs: rewards,
-          unclaimedTokens: unclaimedTokens,
-          userUnclaimedRewards,
-        });
-      } catch (err) {
-        captureException(
-          new Error(
-            `Could not fetch Transactions count for ${address}, on ${network}`
-          ),
-          {
-            tags: {
-              section: "dashboard",
-            },
-          }
-        );
-        return;
-      }
-    })();
-
-    // Get claimed Rewards
-    (async () => {
-      try {
-        const { data, errors } = await useUserRewardsByAddress(
-          network,
-          address
-        );
-
-        if (errors || !data) {
-          throw errors;
-        }
-
-        const { winners } = data;
-
-        const totalRewards = winners.reduce(
-          (sum, { winning_amount, token_decimals }) =>
-            sum + winning_amount / 10 ** token_decimals,
-          0
-        );
-
-        setUserTotalRewards(totalRewards);
-      } catch (err) {
-        captureException(
-          new Error(
-            `Could not fetch Transactions count for ${address}, on ${network}`
-          ),
-          {
-            tags: {
-              section: "dashboard",
-            },
-          }
-        );
-        return;
-      }
-    })();
-  }, [connected, address]);
+  // Get claimed Rewards
 
   const unclaimedRewardColumns = isSmallMobile
     ? [{ name: "TOKEN" }, { name: "REWARD" }]
