+++ web/app.fluidity.money/package.json
@@ -85,6 +85,7 @@
     "@remix-run/node": "^1.7.1",
     "@sentry/react": "^7.12.1",
     "@sentry/tracing": "^7.12.1",
+    "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@types/bn.js": "^5.1.1",
     "@types/compression": "^1.7.2",
@@ -102,12 +103,14 @@
     "@types/tinycolor2": "^1.4.3",
     "@typescript-eslint/eslint-plugin": "^5.37.0",
     "@typescript-eslint/parser": "^5.37.0",
+    "cross-fetch": "^3.1.5",
     "dotenv": "^16.0.3",
     "esbuild": "^0.15.10",
     "eslint": "^8.23.1",
     "eslint-plugin-react": "^7.31.8",
     "graphqurl": "^1.0.1",
     "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
     "npm-run-all": "^4.1.5",
     "remix-esbuild-override": "^3.0.4",
     "ts-jest": "^29.0.3",
